Object loading become more friendly now. The structure of Object3D was changed. Class Face was removed. Module "parser.py" was renamed and moved to folder "geometry"
